rename tof_1d_correlate
